Move logic to apply quorun on the SentinelPool.MonitoredShards() to the redis package
